♻️ FixedRawSquare -> FixedRaw
