+++ python_prototype/src/tqec/templates/fixed/base.py
@@ -1,4 +1,5 @@
 import typing as ty
+import numpy
 from tqec.templates.base import Template
 
 
@@ -12,3 +13,23 @@ class FixedTemplate(Template):
 
     def to_dict(self) -> dict[str, ty.Any]:
         return {"scalable": False}
+
+
+class FixedRaw(FixedTemplate):
+    def __init__(self, plaquette_template: list[list[int]]) -> None:
+        FixedTemplate.__init__(self)
+        self._plaquette_template = plaquette_template
+
+    def to_dict(self) -> dict[str, ty.Any]:
+        ret = FixedTemplate.to_dict(self)
+        ret.update({"plaquette_template": self._plaquette_template})
+        return ret
+
+    def instanciate(self, *plaquette_indices: int) -> numpy.ndarray:
+        return numpy.array(plaquette_indices)[self._plaquette_template]
+
+    @property
+    def shape(self) -> tuple[int, int]:
+        if len(self._plaquette_template) == 0:
+            return (0, 0)
+        return (len(self._plaquette_template), len(self._plaquette_template[0]))
